deprecate worker_done, add worker_finished
Replace the worker_done event with the worker_finished event in the
synopsis and document both, deprecating worker_done.

The difference being that worker_finished fires after the worker has
been removed so that num_workers and has_workers works correctly inside
the event.

Update all tests to reflect this.