more Win32 fixes
Stop using Win32::ShellQuote for external programs, apparently that only
worked from a cygwin shell but not from cmd. The MakeMaker::Awesome hack
also goes away with this.

Automatically around worker_stdout and worker_stderr, if they are
defined, to remove line ends if ::Filter::Line is being used (which is
the default) for external programs.

Skip 10.worker.enqueue.t on Win32 which seems to sometimes fail on win32
smokers. I cannot reproduce this and have no idea why.